Refactor: Convert btree function to class
